Refactor: Organize source/ui directory into subdirectories
Moved window files to source/ui/windows/
Moved dialog files to source/ui/dialogs/
Moved main_menubar to source/ui/menubar/
Moved main_toolbar to source/ui/toolbar/
Updated CMakeLists.txt and include paths accordingly.
