+++ source/app/application.cpp
@@ -27,10 +27,10 @@
 #include "ui/dialogs/goto_position_dialog.h"
 #include "palette/palette_window.h"
 #include "app/preferences.h"
-#include "ui/result_window.h"
+#include "ui/windows/result_window.h"
 #include "rendering/ui/minimap_window.h"
-#include "ui/about_window.h"
-#include "ui/main_menubar.h"
+#include "ui/windows/about_window.h"
+#include "ui/menubar/main_menubar.h"
 #include "app/updater.h"
 #include "ui/artprovider.h"
 #include "ui/map/export_tilesets_window.h"
