+++ source/brushes/managers/brush_manager.cpp
@@ -15,7 +15,7 @@
 #include "map/map.h"
 #include "map/basemap.h"
 #include "ui/gui.h"
-#include "ui/main_toolbar.h"
+#include "ui/toolbar/main_toolbar.h"
 #include <algorithm>
 #include <array>
