+++ source/editor/managers/editor_manager.cpp
@@ -7,13 +7,13 @@
 #include "editor/managers/editor_manager.h"
 #include "ui/gui.h"
 #include "ui/map_tab.h"
-#include "ui/main_menubar.h"
+#include "ui/menubar/main_menubar.h"
 #include "palette/palette_window.h"
 #include "editor/editor.h"
 #include "editor/editor_factory.h"
 #include "editor/action_queue.h"
 #include "ui/dialog_util.h"
-#include "ui/welcome_dialog.h"
+#include "ui/dialogs/welcome_dialog.h"
 #include "app/application.h"
 #include "app/managers/version_manager.h"
 #include "ui/managers/status_manager.h"
