+++ source/ui/dialogs/map_statistics_dialog.cpp
@@ -4,7 +4,7 @@
 
 #include "app/main.h"
 
-#include "ui/map_statistics_dialog.h"
+#include "ui/dialogs/map_statistics_dialog.h"
 #include "ui/gui.h"
 #include "ui/dialog_util.h"
 #include "map/map_statistics.h"
