+++ source/ui/managers/layout_manager.cpp
@@ -12,14 +12,14 @@
 #include "palette/managers/palette_manager.h"
 
 #include "rendering/ui/minimap_window.h"
-#include "ui/main_menubar.h"
-#include "ui/main_toolbar.h"
+#include "ui/menubar/main_menubar.h"
+#include "ui/toolbar/main_toolbar.h"
 #include <wx/display.h>
 #include "ui/managers/minimap_manager.h"
 #include "ingame_preview/ingame_preview_manager.h"
 #include "ingame_preview/ingame_preview_window.h"
 #include "ui/managers/status_manager.h"
-#include "ui/tool_options_window.h"
+#include "ui/windows/tool_options_window.h"
 
 LayoutManager g_layout;
