+++ source/ui/properties/container_properties_window.cpp
@@ -9,7 +9,7 @@
 #include "ui/dialog_util.h"
 #include "ui/properties/property_validator.h"
 #include "app/application.h"
-#include "ui/find_item_window.h"
+#include "ui/windows/find_item_window.h"
 #include "ui/gui_ids.h"
 #include "ui/properties/properties_window.h"
 #include "ui/properties/old_properties_window.h"
