+++ source/ui/windows/result_window.h
